Allow system server to access /data on non-preopt
This allows /data access to system_server on non-preopt
builds to prevent poor performance and ram management on non-preopt
builds. The restriction was introduced in O-MR1 but eliminates the point
of having a dexpreopt flag, if it's only gonna result in poor
performace.

Some build environments may have issues with dex on P, so this is a
small work around to allowing build non-preopt whilst maintaining
stability

Change-Id: I1aae0e9f18f5d48e45d2d0f5f399e308c2eb3287